refactor(desktop): split the Tauri bridge into command modules
desktop/src-tauri/src/lib.rs mixed the Bridge state, 37 command handlers,
DTOs, the poll loop, and the entry point in one 1,483-line file. It now
keeps only the core (Bridge, persist_history, state_signature,
ensure_ready, run(), init_bridge, spawn_poll_loop; ~370 lines) and the
commands live in src/commands/ grouped by concern:

  auth.rs      identity lifecycle (unlock/set-password/status) + settings
  chats.rs     conversations, messages, file sends, transfers, rename/delete
  connect.rs   hosting, peer/relay connections, TOFU confirmation
  contacts.rs  contacts + invite links
  party.rs     Communities: join/leave, channels, DMs, files, parties.json

Pure code motion: command names and behavior unchanged (generate_handler!
now references them by module path), so the React frontend is untouched.
Cross-module items are pub(crate). Docs updated. Workspace tests green
(304), clippy clean.
