+++ desktop/src-tauri/src/commands/mod.rs
@@ -1,0 +1,7 @@
+//! Tauri command handlers, grouped by concern. Wired up in `run()`.
+
+pub(crate) mod auth;
+pub(crate) mod chats;
+pub(crate) mod connect;
+pub(crate) mod contacts;
+pub(crate) mod party;
