+++ desktop/src-tauri/src/lib.rs
@@ -91,50 +91,6 @@ fn state_signature(mgr: &ChatManager) -> u64 {
     sig
 }
 
-// ── DTOs sent to the webview ────────────────────────────────────────────────
-
-#[derive(Serialize)]
-struct AuthStatus {
-    state: &'static str,
-    name: String,
-    fingerprint: String,
-}
-
-#[derive(Serialize)]
-struct ConvSummary {
-    id: String,
-    title: String,
-    last: Option<String>,
-    connected: bool,
-    placeholder: bool,
-    kind: &'static str,
-    transport: &'static str,
-    /// True once the peer's fingerprint has been confirmed (TOFU-verified). The
-    /// UI must not claim "verified" for conversations that are still pending.
-    verified: bool,
-    /// Total message count; the frontend derives unread badges from it (count
-    /// beyond what was on screen when the conversation was last open).
-    messages: usize,
-}
-
-fn kind_str(k: messenger_core::types::ChatKind) -> &'static str {
-    use messenger_core::types::ChatKind::*;
-    match k {
-        Dm => "dm",
-        Group => "group",
-        Channel => "channel",
-    }
-}
-
-fn transport_str(t: messenger_core::types::Transport) -> &'static str {
-    use messenger_core::types::Transport::*;
-    match t {
-        Direct => "direct",
-        Relay => "relay",
-        Server => "server",
-    }
-}
-
 /// Reject any post-auth command while the identity is still locked or a password
 /// setup is pending. The React shell already hides the UI in those states, but
 /// the bridge must enforce the barrier itself so no command can mutate state or
@@ -149,1079 +105,8 @@ fn ensure_ready(state: &Bridge) -> Result<(), String> {
     Ok(())
 }
 
-// ── Commands ────────────────────────────────────────────────────────────────
-
-#[tauri::command]
-fn auth_status(state: tauri::State<'_, Bridge>) -> AuthStatus {
-    let id = state.identity.lock().unwrap();
-    let st = if id.is_locked() {
-        "unlock"
-    } else if *state.is_new.lock().unwrap() || *state.force_setup.lock().unwrap() {
-        "set_password"
-    } else {
-        "ready"
-    };
-    AuthStatus {
-        state: st,
-        name: id.name.clone(),
-        fingerprint: id.fingerprint.clone(),
-    }
-}
-
-/// Unlock an existing, password-protected identity and load history.
-#[tauri::command]
-async fn unlock(password: String, state: tauri::State<'_, Bridge>) -> Result<(), String> {
-    let key = {
-        let mut id = state.identity.lock().unwrap();
-        id.decrypt(&password)
-            .map_err(|_| "Wrong password".to_string())?;
-        let key = id.history_key().map_err(|e| e.to_string())?;
-        *state.is_new.lock().unwrap() = false;
-        *state.force_setup.lock().unwrap() = false;
-        key
-    };
-    let mut mgr = state.manager.lock().await;
-    mgr.set_history_key(key);
-    if let Err(e) = mgr.load_history_auto(&state.history_path, &key) {
-        tracing::warn!("Failed to load history after unlock: {}", e);
-    }
-    auto_host_if_configured(&state, &mut mgr).await;
-    Ok(())
-}
-
-/// Honor `auto_host_on_startup` once the identity is unlocked and history (with
-/// its persisted config) is loaded — "open the app and be reachable", matching
-/// the egui/TUI apps. Failures are logged, never fatal to the unlock.
-async fn auto_host_if_configured(state: &tauri::State<'_, Bridge>, mgr: &mut ChatManager) {
-    if !mgr.config.auto_host_on_startup {
-        return;
-    }
-    let port = mgr.config.listen_port;
-    let pk = { state.identity.lock().unwrap().private_key() };
-    match pk {
-        Ok(pk) => {
-            if let Err(e) = mgr.start_host(port, pk).await {
-                tracing::warn!("auto-host on startup failed: {e}");
-            } else {
-                tracing::info!(port, "auto-hosting on startup");
-            }
-        }
-        Err(e) => tracing::warn!("auto-host: no private key: {e}"),
-    }
-}
-
-/// Set a password on a fresh / plaintext identity, persist it, and stay unlocked.
-#[tauri::command]
-async fn set_password(password: String, state: tauri::State<'_, Bridge>) -> Result<(), String> {
-    let key = {
-        let mut id = state.identity.lock().unwrap();
-        id.encrypt(&password).map_err(|e| e.to_string())?;
-        id.save(&state.identity_save_path())
-            .map_err(|e| e.to_string())?;
-        // Decrypt in memory so we remain unlocked for this session.
-        id.decrypt(&password).map_err(|e| e.to_string())?;
-        let key = id.history_key().map_err(|e| e.to_string())?;
-        *state.is_new.lock().unwrap() = false;
-        *state.force_setup.lock().unwrap() = false;
-        key
-    };
-    // Load any existing history the same way `unlock` does. Without this, the
-    // first post-setup session starts from an empty manager and the next save
-    // could overwrite an existing `history.json.enc` with empty state.
-    let mut mgr = state.manager.lock().await;
-    mgr.set_history_key(key);
-    if let Err(e) = mgr.load_history_auto(&state.history_path, &key) {
-        tracing::warn!("Failed to load history after password setup: {}", e);
-    }
-    auto_host_if_configured(&state, &mut mgr).await;
-    Ok(())
-}
-
-#[tauri::command]
-fn my_identity(state: tauri::State<'_, Bridge>) -> AuthStatus {
-    auth_status(state)
-}
-
-/// The user-facing settings the desktop app exposes. Only fields the core
-/// actually honors are surfaced (a toggle the runtime ignores is a lying UI):
-/// `download_dir` and typing/notification switches are read by `ChatManager`,
-/// and auto-host is implemented by this bridge on unlock.
-#[derive(Serialize)]
-struct SettingsDto {
-    download_dir: String,
-    enable_notifications: bool,
-    enable_typing_indicators: bool,
-    auto_host_on_startup: bool,
-    listen_port: u16,
-}
-
-#[derive(Deserialize)]
-struct SettingsUpdate {
-    enable_notifications: bool,
-    enable_typing_indicators: bool,
-    auto_host_on_startup: bool,
-    listen_port: u16,
-}
-
-#[tauri::command]
-async fn get_settings(state: tauri::State<'_, Bridge>) -> Result<SettingsDto, String> {
-    ensure_ready(&state)?;
-    let mgr = state.manager.lock().await;
-    Ok(SettingsDto {
-        download_dir: mgr.config.download_dir.display().to_string(),
-        enable_notifications: mgr.config.enable_notifications,
-        enable_typing_indicators: mgr.config.enable_typing_indicators,
-        auto_host_on_startup: mgr.config.auto_host_on_startup,
-        listen_port: mgr.config.listen_port,
-    })
-}
-
-/// Apply the toggle/number settings and persist them (config is part of the
-/// encrypted history file). The download directory has its own picker command.
-#[tauri::command]
-async fn update_settings(
-    settings: SettingsUpdate,
-    state: tauri::State<'_, Bridge>,
-) -> Result<(), String> {
-    ensure_ready(&state)?;
-    if settings.listen_port == 0 {
-        return Err("listen port must be between 1 and 65535".to_string());
-    }
-    {
-        let mut mgr = state.manager.lock().await;
-        mgr.config.enable_notifications = settings.enable_notifications;
-        mgr.config.enable_typing_indicators = settings.enable_typing_indicators;
-        mgr.config.auto_host_on_startup = settings.auto_host_on_startup;
-        mgr.config.listen_port = settings.listen_port;
-    }
-    persist_history(&state.manager, &state.history_path).await;
-    Ok(())
-}
-
-/// Pick a new download directory with the native folder dialog (on a blocking
-/// thread) and persist it. Returns the new path, or `None` if cancelled.
-#[tauri::command]
-async fn pick_download_dir(state: tauri::State<'_, Bridge>) -> Result<Option<String>, String> {
-    ensure_ready(&state)?;
-    let picked = tokio::task::spawn_blocking(|| rfd::FileDialog::new().pick_folder())
-        .await
-        .map_err(|e| e.to_string())?;
-    let Some(dir) = picked else {
-        return Ok(None);
-    };
-    {
-        let mut mgr = state.manager.lock().await;
-        mgr.config.download_dir = dir.clone();
-    }
-    persist_history(&state.manager, &state.history_path).await;
-    Ok(Some(dir.display().to_string()))
-}
-
-#[tauri::command]
-async fn list_conversations(state: tauri::State<'_, Bridge>) -> Result<Vec<ConvSummary>, String> {
-    ensure_ready(&state)?;
-    let mgr = state.manager.lock().await;
-    let mut out = Vec::new();
-    for id in mgr.chat_ids() {
-        if let Some(chat) = mgr.get_chat(id) {
-            let last = chat.messages.last().map(|m| match &m.content {
-                MessageContent::Text { text } => text.clone(),
-                MessageContent::File { filename, .. } => format!("📎 {}", filename),
-                MessageContent::Edited { new_text } => new_text.clone(),
-            });
-            out.push(ConvSummary {
-                id: id.to_string(),
-                title: chat.title.clone(),
-                last,
-                connected: mgr.is_connected(&id),
-                placeholder: chat.is_host_placeholder,
-                kind: kind_str(chat.kind),
-                transport: transport_str(chat.transport),
-                verified: chat.peer_fingerprint.is_some(),
-                messages: chat.messages.len(),
-            });
-        }
-    }
-    Ok(out)
-}
-
-/// A live file transfer, for progress display in the chat pane.
-#[derive(Serialize)]
-struct TransferDto {
-    id: String,
-    chat_id: String,
-    filename: String,
-    size: u64,
-    received: u64,
-    status: &'static str,
-    /// Failure reason when `status == "failed"`.
-    error: Option<String>,
-}
-
-fn transfer_status_parts(s: &TransferStatus) -> (&'static str, Option<String>) {
-    match s {
-        TransferStatus::Pending => ("pending", None),
-        TransferStatus::InProgress => ("active", None),
-        TransferStatus::Completed => ("done", None),
-        TransferStatus::Failed(e) => ("failed", Some(e.clone())),
-        TransferStatus::Cancelled => ("cancelled", None),
-    }
-}
-
-/// The active file transfers (both directions), polled by the frontend on the
-/// same `state-updated` cadence as the conversation list, so large sends and
-/// receives show live progress instead of nothing.
-#[tauri::command]
-async fn list_transfers(state: tauri::State<'_, Bridge>) -> Result<Vec<TransferDto>, String> {
-    ensure_ready(&state)?;
-    let mgr = state.manager.lock().await;
-    Ok(mgr
-        .active_transfers_snapshot()
-        .into_iter()
-        .map(|t| {
-            let (status, error) = transfer_status_parts(&t.status);
-            TransferDto {
-                id: t.id.to_string(),
-                chat_id: t.chat_id.to_string(),
-                filename: t.filename,
-                size: t.size,
-                received: t.received,
-                status,
-                error,
-            }
-        })
-        .collect())
-}
-
-/// Return the full conversation (with messages) as JSON for the chat pane.
-#[tauri::command]
-async fn get_conversation(
-    id: String,
-    state: tauri::State<'_, Bridge>,
-) -> Result<serde_json::Value, String> {
-    ensure_ready(&state)?;
-    let uuid = Uuid::parse_str(&id).map_err(|e| e.to_string())?;
-    let mgr = state.manager.lock().await;
-    match mgr.get_chat(uuid) {
-        Some(chat) => serde_json::to_value(chat).map_err(|e| e.to_string()),
-        None => Err("No such conversation".to_string()),
-    }
-}
-
-#[tauri::command]
-async fn send_message(
-    id: String,
-    text: String,
-    state: tauri::State<'_, Bridge>,
-) -> Result<(), String> {
-    ensure_ready(&state)?;
-    let uuid = Uuid::parse_str(&id).map_err(|e| e.to_string())?;
-    state
-        .manager
-        .lock()
-        .await
-        .send_message(uuid, text)
-        .map_err(|e| e.to_string())?;
-    persist_history(&state.manager, &state.history_path).await;
-    Ok(())
-}
-
-/// Pick a file with the native dialog and send it over the given conversation.
-/// The picker runs on a blocking thread so it never stalls the async runtime;
-/// a cancelled dialog is a successful no-op.
-#[tauri::command]
-async fn send_file(id: String, state: tauri::State<'_, Bridge>) -> Result<(), String> {
-    ensure_ready(&state)?;
-    let uuid = Uuid::parse_str(&id).map_err(|e| e.to_string())?;
-    let picked = tokio::task::spawn_blocking(|| rfd::FileDialog::new().pick_file())
-        .await
-        .map_err(|e| e.to_string())?;
-    let Some(path) = picked else {
-        return Ok(()); // user cancelled
-    };
-    state
-        .manager
-        .lock()
-        .await
-        .send_file(uuid, path)
-        .await
-        .map_err(|e| e.to_string())?;
-    persist_history(&state.manager, &state.history_path).await;
-    Ok(())
-}
-
-#[tauri::command]
-async fn start_host(port: u16, state: tauri::State<'_, Bridge>) -> Result<(), String> {
-    ensure_ready(&state)?;
-    let pk = {
-        let id = state.identity.lock().unwrap();
-        id.private_key().map_err(|e| e.to_string())?
-    };
-    state
-        .manager
-        .lock()
-        .await
-        .start_host(port, pk)
-        .await
-        .map(|_chat_id| ())
-        .map_err(|e| e.to_string())
-}
-
-#[tauri::command]
-async fn connect_peer(
-    host: String,
-    port: u16,
-    state: tauri::State<'_, Bridge>,
-) -> Result<(), String> {
-    ensure_ready(&state)?;
-    let pk = {
-        let id = state.identity.lock().unwrap();
-        id.private_key().map_err(|e| e.to_string())?
-    };
-    state
-        .manager
-        .lock()
-        .await
-        .connect_to_host(&host, port, None, pk)
-        .await
-        .map(|_chat_id| ())
-        .map_err(|e| e.to_string())
-}
-
-/// Start hosting through a blind relay broker. Returns the connection token the
-/// peer needs (alongside the relay address) to dial in — relay is a *transport*,
-/// the conversation is still a verified DM.
-#[tauri::command]
-async fn host_via_relay(relay: String, state: tauri::State<'_, Bridge>) -> Result<String, String> {
-    ensure_ready(&state)?;
-    let pk = {
-        let id = state.identity.lock().unwrap();
-        id.private_key().map_err(|e| e.to_string())?
-    };
-    state
-        .manager
-        .lock()
-        .await
-        .start_host_via_relay(&relay, None, pk)
-        .await
-        .map(|(_chat_id, token)| token)
-        .map_err(|e| e.to_string())
-}
-
-/// Dial a peer through a relay using the relay address + the token they shared.
-#[tauri::command]
-async fn connect_via_relay(
-    relay: String,
-    token: String,
-    state: tauri::State<'_, Bridge>,
-) -> Result<(), String> {
-    ensure_ready(&state)?;
-    let pk = {
-        let id = state.identity.lock().unwrap();
-        id.private_key().map_err(|e| e.to_string())?
-    };
-    state
-        .manager
-        .lock()
-        .await
-        .connect_via_relay(&relay, &token, None, pk)
-        .await
-        .map(|_chat_id| ())
-        .map_err(|e| e.to_string())
-}
-
-#[tauri::command]
-async fn confirm_fingerprint(
-    id: String,
-    accept: bool,
-    state: tauri::State<'_, Bridge>,
-) -> Result<(), String> {
-    ensure_ready(&state)?;
-    let uuid = Uuid::parse_str(&id).map_err(|e| e.to_string())?;
-    let result = state
-        .manager
-        .lock()
-        .await
-        .confirm_fingerprint(uuid, accept)
-        .map_err(|e| e.to_string());
-    // Only clear the pending prompt on success. If confirmation failed, keep it so
-    // `pending_fingerprint()` can still surface it and the user can retry.
-    if result.is_ok() {
-        *state.pending_fp.lock().unwrap() = None;
-        // An accepted fingerprint is persisted onto the chat; save so trust survives a restart.
-        persist_history(&state.manager, &state.history_path).await;
-    }
-    result
-}
-
-/// The pending TOFU fingerprint prompt, if any. The frontend polls this so a
-/// dropped `fingerprint-request` event never leaves a session stuck unverified.
-#[tauri::command]
-fn pending_fingerprint(state: tauri::State<'_, Bridge>) -> Option<serde_json::Value> {
-    if ensure_ready(&state).is_err() {
-        return None;
-    }
-    state
-        .pending_fp
-        .lock()
-        .unwrap()
-        .clone()
-        .map(|(fingerprint, peer_name, chat_id)| {
-            serde_json::json!({
-                "fingerprint": fingerprint,
-                "peer_name": peer_name,
-                "chat_id": chat_id,
-            })
-        })
-}
-
-#[tauri::command]
-async fn rename_chat(
-    id: String,
-    title: String,
-    state: tauri::State<'_, Bridge>,
-) -> Result<(), String> {
-    ensure_ready(&state)?;
-    let uuid = Uuid::parse_str(&id).map_err(|e| e.to_string())?;
-    state
-        .manager
-        .lock()
-        .await
-        .rename_chat(uuid, title)
-        .map_err(|e| e.to_string())?;
-    persist_history(&state.manager, &state.history_path).await;
-    Ok(())
-}
-
-#[tauri::command]
-async fn delete_chat(id: String, state: tauri::State<'_, Bridge>) -> Result<(), String> {
-    ensure_ready(&state)?;
-    let uuid = Uuid::parse_str(&id).map_err(|e| e.to_string())?;
-    state.manager.lock().await.delete_chat(uuid);
-    persist_history(&state.manager, &state.history_path).await;
-    Ok(())
-}
-
-// ── Contacts + invite links ─────────────────────────────────────────────────
-
-#[derive(Serialize)]
-struct ContactDto {
-    id: String,
-    name: String,
-    fingerprint: Option<String>,
-    address: Option<String>,
-    trust: &'static str,
-}
-
-fn trust_str(t: &messenger_core::types::TrustState) -> &'static str {
-    use messenger_core::types::TrustState::*;
-    match t {
-        Unverified => "unverified",
-        Verified => "verified",
-        Trusted => "trusted",
-        Blocked => "blocked",
-    }
-}
-
-fn contact_dto(c: &messenger_core::types::Contact) -> ContactDto {
-    ContactDto {
-        id: c.id.to_string(),
-        name: c.name.clone(),
-        fingerprint: c.fingerprint.clone(),
-        address: c.address.clone(),
-        trust: trust_str(&c.trust_state),
-    }
-}
-
-#[tauri::command]
-async fn list_contacts(state: tauri::State<'_, Bridge>) -> Result<Vec<ContactDto>, String> {
-    ensure_ready(&state)?;
-    let mgr = state.manager.lock().await;
-    Ok(mgr.contacts.values().map(contact_dto).collect())
-}
-
-/// The current user's signed invite link (address derived from the local IP +
-/// the configured listen port, when resolvable).
-#[tauri::command]
-async fn my_invite_link(state: tauri::State<'_, Bridge>) -> Result<String, String> {
-    ensure_ready(&state)?;
-    let address = {
-        let mgr = state.manager.lock().await;
-        let port = mgr.config.listen_port;
-        messenger_core::util::primary_local_ipv4()
-            .map(|ip| messenger_core::util::format_host_port(&ip, port))
-    };
-    let id = state.identity.lock().unwrap();
-    id.generate_signed_invite_link(address)
-        .map_err(|e| e.to_string())
-}
-
-/// Parse an invite link and store it as a contact.
-#[tauri::command]
-async fn import_invite(
-    link: String,
-    state: tauri::State<'_, Bridge>,
-) -> Result<ContactDto, String> {
-    ensure_ready(&state)?;
-    let contact = {
-        let mgr = state.manager.lock().await;
-        mgr.parse_invite_link(&link).map_err(|e| e.to_string())?
-    };
-    let mut mgr = state.manager.lock().await;
-    let id = mgr.import_contact(contact);
-    Ok(contact_dto(mgr.get_contact(id).expect("just inserted")))
-}
-
-/// Dial a stored contact by its saved address.
-#[tauri::command]
-async fn connect_contact(id: String, state: tauri::State<'_, Bridge>) -> Result<(), String> {
-    ensure_ready(&state)?;
-    let uuid = Uuid::parse_str(&id).map_err(|e| e.to_string())?;
-    let address = {
-        let mgr = state.manager.lock().await;
-        mgr.get_contact(uuid).and_then(|c| c.address.clone())
-    };
-    let address = address.ok_or_else(|| "Contact has no saved address".to_string())?;
-    let (host, port) =
-        messenger_core::util::parse_host_port(&address, Some(messenger_core::PORT_DEFAULT))
-            .map_err(|e| e.to_string())?;
-    let pk = {
-        let id = state.identity.lock().unwrap();
-        id.private_key().map_err(|e| e.to_string())?
-    };
-    state
-        .manager
-        .lock()
-        .await
-        .connect_to_host(&host, port, None, pk)
-        .await
-        .map(|_chat_id| ())
-        .map_err(|e| e.to_string())
-}
-
-// ── Communities (Party servers) ─────────────────────────────────────────────
-//
-// A thin bridge over `PartyManager`, mirroring the egui Party tab. Command
-// params are single words (`server`, `channel`, …) to avoid the Tauri 2
-// arg-naming footgun where a snake_case param silently no-ops.
-
-#[derive(Serialize)]
-struct PartyMemberDto {
-    id: String,
-    username: String,
-    online: bool,
-    is_me: bool,
-    /// Message count in my DM thread with this member (0 when not joined yet).
-    /// The frontend derives DM unread badges from it.
-    dm_messages: usize,
-}
-
-#[derive(Serialize)]
-struct PartyChannelDto {
-    id: String,
-    name: String,
-    /// Message count in this channel; the frontend derives unread badges from it.
-    messages: usize,
-}
-
-#[derive(Serialize)]
-struct PartyServerDto {
-    id: String,
-    name: String,
-    address: String,
-    /// The username this client joined with (for rejoin flows).
-    username: String,
-    fingerprint: String,
-    status: &'static str,
-    status_detail: Option<String>,
-    member_id: Option<String>,
-    channels: Vec<PartyChannelDto>,
-    members: Vec<PartyMemberDto>,
-    last_error: Option<String>,
-}
-
-#[derive(Serialize)]
-struct PartyMessageDto {
-    sender_name: String,
-    from_me: bool,
-    kind: &'static str,
-    text: String,
-    size: Option<u64>,
-    /// Content hash of a file message, used to request its download. `None` for text.
-    hash: Option<String>,
-    timestamp: u64,
-}
-
-fn party_status_parts(status: &PartyStatus) -> (&'static str, Option<String>) {
-    match status {
-        PartyStatus::Connecting => ("connecting", None),
-        PartyStatus::Joined => ("joined", None),
-        PartyStatus::Rejected(reason) => ("rejected", Some(reason.clone())),
-        PartyStatus::Disconnected => ("disconnected", None),
-    }
-}
-
-/// Serialize one connection's directory (channels + members), resolving "you".
-fn server_dto(
-    id: Uuid,
-    conn: &encodeur_rsa_rust::app::party_manager::PartyServerConn,
-) -> PartyServerDto {
-    let (status, status_detail) = party_status_parts(&conn.status);
-    let members = conn
-        .members
-        .iter()
-        .map(|m| PartyMemberDto {
-            id: m.id.to_string(),
-            username: m.username.clone(),
-            online: m.online,
-            is_me: conn.member_id == Some(m.id),
-            dm_messages: conn
-                .member_id
-                .map(|me| {
-                    let thread = messenger_core::party::dm_thread_id(me, m.id);
-                    conn.messages.get(&thread).map_or(0, |v| v.len())
-                })
-                .unwrap_or(0),
-        })
-        .collect();
-    let channels = conn
-        .channels
-        .iter()
-        .map(|c| PartyChannelDto {
-            id: c.id.to_string(),
-            name: c.name.clone(),
-            messages: conn.messages.get(&c.id).map_or(0, |v| v.len()),
-        })
-        .collect();
-    PartyServerDto {
-        id: id.to_string(),
-        name: conn.server_name.clone(),
-        address: conn.address.clone(),
-        username: conn.username.clone(),
-        fingerprint: conn.server_fingerprint.clone(),
-        status,
-        status_detail,
-        member_id: conn.member_id.map(|m| m.to_string()),
-        channels,
-        members,
-        last_error: conn.last_error.clone(),
-    }
-}
-
-/// Turn a stored envelope into a display DTO, resolving the sender's username
-/// from the member directory and flagging the local user's own messages.
-fn message_dto(
-    env: &messenger_core::party::Envelope,
-    conn: &encodeur_rsa_rust::app::party_manager::PartyServerConn,
-) -> PartyMessageDto {
-    use messenger_core::party::MessagePayload;
-    let sender_name = conn
-        .members
-        .iter()
-        .find(|m| m.id == env.sender)
-        .map(|m| m.username.clone())
-        .unwrap_or_else(|| "unknown".to_string());
-    let (kind, text, size, hash) = match &env.payload {
-        MessagePayload::Text(t) => ("text", t.clone(), None, None),
-        MessagePayload::File(f) => ("file", f.name.clone(), Some(f.size), Some(f.hash.clone())),
-    };
-    PartyMessageDto {
-        sender_name,
-        from_me: conn.member_id == Some(env.sender),
-        kind,
-        text,
-        size,
-        hash,
-        timestamp: env.timestamp,
-    }
-}
-
-/// Connect to a community server, verify (TOFU) its fingerprint out of band, and
-/// join with a username. Returns the local server id.
-#[tauri::command]
-async fn party_join(
-    address: String,
-    username: String,
-    password: String,
-    state: tauri::State<'_, Bridge>,
-) -> Result<String, String> {
-    ensure_ready(&state)?;
-    let address = address.trim().to_string();
-    let username = username.trim().to_string();
-    let pk = {
-        let id = state.identity.lock().unwrap();
-        id.private_key().map_err(|e| e.to_string())?
-    };
-    let password = Some(password).filter(|p| !p.trim().is_empty());
-    let mut party = state.party.lock().await;
-    let sid = party
-        .connect_and_join(&address, &username, password, &pk)
-        .await
-        .map_err(|e| e.to_string())?;
-
-    // TOFU pinning: if we've joined this address before, its identity must not
-    // have changed. A mismatch is either a redeployed server or an active MITM —
-    // refuse and tell the user, never silently trust the new key.
-    let fingerprint = party
-        .server(sid)
-        .map(|c| c.server_fingerprint.clone())
-        .unwrap_or_default();
-    let saved = load_saved_parties(&state.parties_path);
-    if let Some(pinned) = saved
-        .iter()
-        .find(|p| p.address.eq_ignore_ascii_case(&address) && !p.fingerprint.is_empty())
-    {
-        if pinned.fingerprint != fingerprint {
-            party.remove_server(sid);
-            return Err(format!(
-                "SECURITY: this server's identity changed since you last joined \
-                 (expected {}…, got {}…). If the operator redeployed the server this \
-                 may be expected — leave the saved community and rejoin to trust the \
-                 new identity. Otherwise, do not proceed.",
-                &pinned.fingerprint[..16.min(pinned.fingerprint.len())],
-                &fingerprint[..16.min(fingerprint.len())]
-            ));
-        }
-    }
-    // Rejoin replaces: drop any older entry for the same address (e.g. a
-    // disconnected or join-rejected zombie) so the list never shows duplicates.
-    let stale: Vec<Uuid> = party
-        .server_ids()
-        .into_iter()
-        .filter(|id| {
-            *id != sid
-                && party
-                    .server(*id)
-                    .is_some_and(|c| c.address.eq_ignore_ascii_case(&address))
-        })
-        .collect();
-    for id in stale {
-        party.remove_server(id);
-    }
-    drop(party);
-
-    upsert_saved_party(
-        &state.parties_path,
-        SavedParty {
-            address,
-            username,
-            name: String::new(),
-            fingerprint,
-        },
-    );
-    Ok(sid.to_string())
-}
-
-/// The joined community servers with their channels and member directories.
-#[tauri::command]
-async fn party_list(state: tauri::State<'_, Bridge>) -> Result<Vec<PartyServerDto>, String> {
-    ensure_ready(&state)?;
-    let party = state.party.lock().await;
-    let mut out: Vec<PartyServerDto> = party
-        .server_ids()
-        .into_iter()
-        .filter_map(|id| party.server(id).map(|conn| server_dto(id, conn)))
-        .collect();
-    // Stable order so the UI's server list doesn't reshuffle each poll.
-    out.sort_by(|a, b| a.name.cmp(&b.name).then(a.id.cmp(&b.id)));
-    Ok(out)
-}
-
-/// Messages in a channel (already seeded from durable history on join; live
-/// posts arrive via the poll loop).
-#[tauri::command]
-async fn party_history(
-    server: String,
-    channel: String,
-    state: tauri::State<'_, Bridge>,
-) -> Result<Vec<PartyMessageDto>, String> {
-    ensure_ready(&state)?;
-    let sid = Uuid::parse_str(&server).map_err(|e| e.to_string())?;
-    let cid = Uuid::parse_str(&channel).map_err(|e| e.to_string())?;
-    let party = state.party.lock().await;
-    let conn = party
-        .server(sid)
-        .ok_or_else(|| "unknown server".to_string())?;
-    let msgs = conn
-        .messages
-        .get(&cid)
-        .map(|v| v.iter().map(|e| message_dto(e, conn)).collect())
-        .unwrap_or_default();
-    Ok(msgs)
-}
-
-/// Post a text message to a channel.
-#[tauri::command]
-async fn party_post(
-    server: String,
-    channel: String,
-    text: String,
-    state: tauri::State<'_, Bridge>,
-) -> Result<(), String> {
-    ensure_ready(&state)?;
-    let sid = Uuid::parse_str(&server).map_err(|e| e.to_string())?;
-    let cid = Uuid::parse_str(&channel).map_err(|e| e.to_string())?;
-    state
-        .party
-        .lock()
-        .await
-        .post(sid, cid, text)
-        .map_err(|e| e.to_string())
-}
-
-/// Create a new channel on a community server.
-#[tauri::command]
-async fn party_create_channel(
-    server: String,
-    name: String,
-    state: tauri::State<'_, Bridge>,
-) -> Result<(), String> {
-    ensure_ready(&state)?;
-    let sid = Uuid::parse_str(&server).map_err(|e| e.to_string())?;
-    state
-        .party
-        .lock()
-        .await
-        .create_channel(sid, name)
-        .map_err(|e| e.to_string())
-}
-
-/// Send a direct message to another member of a community server.
-#[tauri::command]
-async fn party_send_dm(
-    server: String,
-    to: String,
-    text: String,
-    state: tauri::State<'_, Bridge>,
-) -> Result<(), String> {
-    ensure_ready(&state)?;
-    let sid = Uuid::parse_str(&server).map_err(|e| e.to_string())?;
-    let to = Uuid::parse_str(&to).map_err(|e| e.to_string())?;
-    state
-        .party
-        .lock()
-        .await
-        .send_dm(sid, to, text)
-        .map_err(|e| e.to_string())
-}
-
-/// The DM thread with a member. Requests a fresh fetch (offline catch-up); the
-/// authoritative history lands on a later poll.
-#[tauri::command]
-async fn party_dm_history(
-    server: String,
-    peer: String,
-    state: tauri::State<'_, Bridge>,
-) -> Result<Vec<PartyMessageDto>, String> {
-    ensure_ready(&state)?;
-    let sid = Uuid::parse_str(&server).map_err(|e| e.to_string())?;
-    let peer = Uuid::parse_str(&peer).map_err(|e| e.to_string())?;
-    let party = state.party.lock().await;
-    let _ = party.fetch_dm_history(sid, peer);
-    let conn = party
-        .server(sid)
-        .ok_or_else(|| "unknown server".to_string())?;
-    Ok(party
-        .dm_messages(sid, peer)
-        .iter()
-        .map(|e| message_dto(e, conn))
-        .collect())
-}
-
-/// Clear a server's last surfaced error (e.g. a rejected post) after the UI has
-/// shown it.
-#[tauri::command]
-async fn party_clear_error(server: String, state: tauri::State<'_, Bridge>) -> Result<(), String> {
-    ensure_ready(&state)?;
-    let sid = Uuid::parse_str(&server).map_err(|e| e.to_string())?;
-    state.party.lock().await.clear_server_error(sid);
-    Ok(())
-}
-
-/// A community this client has joined, persisted (plaintext, no secrets — the
-/// password is never stored) so it survives restarts: the UI offers one-click
-/// rejoin, and the pinned server fingerprint turns the first join's TOFU into a
-/// real trust anchor — a different identity at the same address is refused.
-#[derive(Clone, Serialize, Deserialize)]
-struct SavedParty {
-    address: String,
-    username: String,
-    #[serde(default)]
-    name: String,
-    #[serde(default)]
-    fingerprint: String,
-}
-
-/// Load the saved-communities list; missing or unreadable files are an empty list
-/// (never an error — this is convenience state, not critical data).
-fn load_saved_parties(path: &Path) -> Vec<SavedParty> {
-    std::fs::read_to_string(path)
-        .ok()
-        .and_then(|s| serde_json::from_str(&s).ok())
-        .unwrap_or_default()
-}
-
-/// Best-effort save of the saved-communities list; errors are logged, not fatal.
-fn save_saved_parties(path: &Path, list: &[SavedParty]) {
-    match serde_json::to_string_pretty(list) {
-        Ok(json) => {
-            if let Err(e) = std::fs::write(path, json) {
-                tracing::warn!("saving communities list failed: {e}");
-            }
-        }
-        Err(e) => tracing::warn!("serializing communities list failed: {e}"),
-    }
-}
-
-/// Insert or update the entry for `address` (matched case-insensitively).
-fn upsert_saved_party(path: &Path, entry: SavedParty) {
-    let mut list = load_saved_parties(path);
-    match list
-        .iter_mut()
-        .find(|p| p.address.eq_ignore_ascii_case(&entry.address))
-    {
-        Some(existing) => {
-            existing.username = entry.username;
-            if !entry.name.is_empty() {
-                existing.name = entry.name;
-            }
-            if !entry.fingerprint.is_empty() {
-                existing.fingerprint = entry.fingerprint;
-            }
-        }
-        None => list.push(entry),
-    }
-    save_saved_parties(path, &list);
-}
-
-/// The saved communities, for the join screen's one-click rejoin list.
-#[tauri::command]
-async fn party_saved(state: tauri::State<'_, Bridge>) -> Result<Vec<SavedParty>, String> {
-    ensure_ready(&state)?;
-    Ok(load_saved_parties(&state.parties_path))
-}
-
-/// Leave a community: drop the connection, forget its local state, and remove it
-/// from the saved list. The server keeps the membership, so rejoining with the
-/// same identity later resumes it.
-#[tauri::command]
-async fn party_leave(server: String, state: tauri::State<'_, Bridge>) -> Result<(), String> {
-    ensure_ready(&state)?;
-    let sid = Uuid::parse_str(&server).map_err(|e| e.to_string())?;
-    let mut party = state.party.lock().await;
-    let address = party.server(sid).map(|c| c.address.clone());
-    if !party.remove_server(sid) {
-        return Err("unknown server".to_string());
-    }
-    drop(party);
-    if let Some(addr) = address {
-        let mut list = load_saved_parties(&state.parties_path);
-        list.retain(|p| !p.address.eq_ignore_ascii_case(&addr));
-        save_saved_parties(&state.parties_path, &list);
-    }
-    Ok(())
-}
-
-/// Open the native file picker (on a blocking thread) and read the chosen file's
-/// bytes, returning `(name, mime, data)`. `Ok(None)` if the user cancelled.
-async fn pick_upload() -> Result<Option<(String, String, Vec<u8>)>, String> {
-    let picked = tokio::task::spawn_blocking(|| rfd::FileDialog::new().pick_file())
-        .await
-        .map_err(|e| e.to_string())?;
-    let Some(path) = picked else {
-        return Ok(None);
-    };
-    let name = path
-        .file_name()
-        .map(|n| n.to_string_lossy().to_string())
-        .unwrap_or_else(|| "file".to_string());
-    let mime = messenger_core::util::guess_mime(&path).to_string();
-    let data = std::fs::read(&path).map_err(|e| e.to_string())?;
-    Ok(Some((name, mime, data)))
-}
-
-/// Pick a file and post it to a community channel. The picker runs on a blocking
-/// thread; a cancelled dialog is a successful no-op.
-#[tauri::command]
-async fn party_send_file(
-    server: String,
-    channel: String,
-    state: tauri::State<'_, Bridge>,
-) -> Result<(), String> {
-    ensure_ready(&state)?;
-    let sid = Uuid::parse_str(&server).map_err(|e| e.to_string())?;
-    let cid = Uuid::parse_str(&channel).map_err(|e| e.to_string())?;
-    let Some((name, mime, data)) = pick_upload().await? else {
-        return Ok(());
-    };
-    state
-        .party
-        .lock()
-        .await
-        .send_file(sid, cid, name, mime, data)
-        .map_err(|e| e.to_string())
-}
-
-/// Pick a file and send it as a direct message to another community member.
-#[tauri::command]
-async fn party_send_file_dm(
-    server: String,
-    to: String,
-    state: tauri::State<'_, Bridge>,
-) -> Result<(), String> {
-    ensure_ready(&state)?;
-    let sid = Uuid::parse_str(&server).map_err(|e| e.to_string())?;
-    let peer = Uuid::parse_str(&to).map_err(|e| e.to_string())?;
-    let Some((name, mime, data)) = pick_upload().await? else {
-        return Ok(());
-    };
-    state
-        .party
-        .lock()
-        .await
-        .send_file_dm(sid, peer, name, mime, data)
-        .map_err(|e| e.to_string())
-}
-
-/// Download a community file by content hash and save it via the native dialog.
-/// `name` seeds the save dialog. The server only returns bytes the caller is
-/// allowed to see (access-checked there). A cancelled save dialog is a no-op.
-#[tauri::command]
-async fn party_download_file(
-    server: String,
-    hash: String,
-    name: String,
-    state: tauri::State<'_, Bridge>,
-) -> Result<(), String> {
-    ensure_ready(&state)?;
-    let sid = Uuid::parse_str(&server).map_err(|e| e.to_string())?;
-    // Register the download and drop the party lock BEFORE awaiting, so the poll
-    // loop can lock the manager, drain the FileData response, and complete us.
-    let rx = state
-        .party
-        .lock()
-        .await
-        .request_download(sid, hash)
-        .map_err(|e| e.to_string())?;
-    // Outer error: the sender was dropped (connection torn down). Inner error: the
-    // server refused (file gone / not permitted). Either way, surface a message
-    // rather than hang.
-    let data = match rx.await {
-        Ok(Ok(data)) => data,
-        Ok(Err(reason)) => return Err(reason),
-        Err(_) => return Err("download failed: the connection closed".to_string()),
-    };
-    let suggested = name;
-    let picked = tokio::task::spawn_blocking(move || {
-        rfd::FileDialog::new().set_file_name(suggested).save_file()
-    })
-    .await
-    .map_err(|e| e.to_string())?;
-    let Some(path) = picked else {
-        return Ok(()); // user cancelled the save dialog
-    };
-    std::fs::write(&path, &data).map_err(|e| e.to_string())
-}
+mod commands;
+use commands::party::{upsert_saved_party, SavedParty};
 
 // ── Entry point ─────────────────────────────────────────────────────────────
 
@@ -1257,43 +142,43 @@ pub fn run() {
     tauri::Builder::default()
         .manage(bridge)
         .invoke_handler(tauri::generate_handler![
-            auth_status,
-            unlock,
-            set_password,
-            my_identity,
-            list_conversations,
-            get_conversation,
-            list_transfers,
-            get_settings,
-            update_settings,
-            pick_download_dir,
-            send_message,
-            send_file,
-            start_host,
-            connect_peer,
-            host_via_relay,
-            connect_via_relay,
-            confirm_fingerprint,
-            rename_chat,
-            delete_chat,
-            list_contacts,
-            my_invite_link,
-            import_invite,
-            connect_contact,
-            pending_fingerprint,
-            party_join,
-            party_list,
-            party_history,
-            party_post,
-            party_create_channel,
-            party_send_dm,
-            party_dm_history,
-            party_clear_error,
-            party_send_file,
-            party_send_file_dm,
-            party_download_file,
-            party_saved,
-            party_leave,
+            commands::auth::auth_status,
+            commands::auth::unlock,
+            commands::auth::set_password,
+            commands::auth::my_identity,
+            commands::chats::list_conversations,
+            commands::chats::get_conversation,
+            commands::chats::list_transfers,
+            commands::auth::get_settings,
+            commands::auth::update_settings,
+            commands::auth::pick_download_dir,
+            commands::chats::send_message,
+            commands::chats::send_file,
+            commands::connect::start_host,
+            commands::connect::connect_peer,
+            commands::connect::host_via_relay,
+            commands::connect::connect_via_relay,
+            commands::connect::confirm_fingerprint,
+            commands::chats::rename_chat,
+            commands::chats::delete_chat,
+            commands::contacts::list_contacts,
+            commands::contacts::my_invite_link,
+            commands::contacts::import_invite,
+            commands::contacts::connect_contact,
+            commands::connect::pending_fingerprint,
+            commands::party::party_join,
+            commands::party::party_list,
+            commands::party::party_history,
+            commands::party::party_post,
+            commands::party::party_create_channel,
+            commands::party::party_send_dm,
+            commands::party::party_dm_history,
+            commands::party::party_clear_error,
+            commands::party::party_send_file,
+            commands::party::party_send_file_dm,
+            commands::party::party_download_file,
+            commands::party::party_saved,
+            commands::party::party_leave,
         ])
         .setup(|app| {
             let b = app.state::<Bridge>();
